LPS-93661 Merge test bundle activators.
