+++ modules/apps/portal-security/portal-security-auth-verifier-test/bnd-guest-allowed.bnd
@@ -1,4 +1,4 @@
-Bundle-Activator: com.liferay.portal.security.auth.verifier.test.internal.activator.GuestAllowedAuthVerifierBundleActivator
+Bundle-Activator: com.liferay.portal.security.auth.verifier.test.internal.activator.AuthVerifierBundleActivator
 Bundle-Name: Liferay Portal Security Auth Verifier Test Guest Allowed
 Bundle-SymbolicName: com.liferay.portal.security.auth.verifier.test.guest.allowed
 Bundle-Version: 1.0.0
