+++ modules/apps/portal-security/portal-security-auth-verifier-test/bnd-tracker.bnd
@@ -1,4 +1,4 @@
-Bundle-Activator: com.liferay.portal.security.auth.verifier.test.internal.activator.TrackerAuthVerifierBundleActivator
+Bundle-Activator: com.liferay.portal.security.auth.verifier.test.internal.activator.AuthVerifierBundleActivator
 Bundle-Name: Liferay Portal Security Auth Verifier Test Tracker
 Bundle-SymbolicName: com.liferay.portal.security.auth.verifier.test.tracker
 Bundle-Version: 1.0.0
